Added mobile responsiveness and lifted ledgerHistory state to the parent
